# docs/openapi.yaml
openapi: 3.0.3
info:
  title: WealthOS Backend API
  version: 2.0.0
  description: Swagger documentation for the WealthOS backend.
servers:
  - url: http://localhost:3000
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AcceptInviteRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
...
paths:
...
  /invitations/{token}/accept:
    get:
      tags: [Invitations]
      summary: Invitation accept page (browser)
      description: Devuelve una página HTML con botón para aceptar. Útil cuando el invitado hace clic en el email desde el navegador.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
        - in: query
          name: email
          required: true
          schema: { type: string, format: email }
      responses:
        "200":
          description: HTML page
          content:
            text/html:
              schema:
                type: string
    post:
      tags: [Invitations]
      summary: Accept invitation
      description: |
        Acepta una invitación pendiente. El usuario debe tener cuenta registrada con ese email.
        Agrega al usuario como miembro del espacio compartido del presupuesto.
      parameters:
        - in: path
          name: token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AcceptInviteRequest"
      responses:
        "200": { description: OK }
...